pyproject.toml
[tool.poetry]
name = "observability-project"
version = "0.1.0"
description = ""
authors = ["Rafael Panegassi <rm.panegassi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.10.2"
requests = "^2.32.3"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
opentelemetry-api = "^1.28.2"
opentelemetry-sdk = "^1.28.2"
flask = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
